requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
Pillow==9.4.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
